Move delete code & impl broad index updates
