+++ src/executor/alter_row/mod.rs
@@ -1,11 +1,13 @@
 mod auto_increment;
 mod insert;
+mod delete;
 mod update;
 mod validate;
 mod validate_unique;
 
 pub use {
 	insert::insert,
+	delete::delete,
 	update::update,
 	validate::{columns_to_positions, validate, ValidateError},
 	validate_unique::validate_unique,
