+++ src/executor/execute.rs
@@ -1,18 +1,15 @@
 use {
 	super::{
-		alter_row::{insert, update},
+		alter_row::{insert, update, delete},
 		alter_table::{create_index, create_table, drop, truncate},
 		query::query,
-		types::ColumnInfo,
 	},
 	crate::{
-		data::{get_name, Schema},
 		glue::Context,
-		parse_sql::Query,
-		MetaRecipe, PlannedRecipe, Result, Row, StorageInner, Value,
+		parse_sql::Query, Result, Row, StorageInner, Value,
 	},
 	serde::Serialize,
-	sqlparser::ast::{ColumnDef, SetVariableValue, Statement},
+	sqlparser::ast::{SetVariableValue, Statement},
 	std::convert::TryInto,
 	thiserror::Error as ThisError,
 };
@@ -120,61 +117,7 @@ pub async fn execute(
 		Statement::Delete {
 			table_name,
 			selection,
-		} => {
-			let table_name = get_name(&table_name)?;
-			let Schema { column_defs, .. } = storages[0]
-				.1
-				.fetch_schema(table_name)
-				.await?
-				.ok_or(ExecuteError::TableNotExists)?;
-
-			let columns = column_defs
-				.clone()
-				.into_iter()
-				.map(|column_def| {
-					let ColumnDef { name, .. } = column_def;
-					ColumnInfo::of_name(name.value)
-				})
-				.collect();
-			let filter = selection
-				.clone()
-				.map(|selection| {
-					PlannedRecipe::new(
-						MetaRecipe::new(selection)?.simplify_by_context(context)?,
-						&columns,
-					)
-				})
-				.unwrap_or(Ok(PlannedRecipe::TRUE))?;
-
-			let keys = storages[0]
-				.1
-				.scan_data(table_name)
-				.await?
-				.filter_map(|row_result| {
-					let (key, row) = match row_result {
-						Ok(keyed_row) => keyed_row,
-						Err(error) => return Some(Err(error)),
-					};
-
-					let row = row.0;
-
-					let confirm_constraint = filter.confirm_constraint(&row.clone());
-					match confirm_constraint {
-						Ok(true) => Some(Ok(key)),
-						Ok(false) => None,
-						Err(error) => Some(Err(error)),
-					}
-				})
-				.collect::<Result<Vec<Value>>>()?;
-
-			let num_keys = keys.len();
-
-			storages[0]
-				.1
-				.delete_data(keys)
-				.await
-				.map(|_| Payload::Delete(num_keys))
-		}
+		} => delete(&mut storages, context, table_name, selection).await,
 
 		//- Selection
 		Statement::Query(query_value) => {
